Improve track handling and update FATFS config
Track navigation now wraps based on the actual number of tracks in the audio directory, determined at initialization. The decode time function's return type was corrected to uint16_t. FATFS configuration was updated to use code page 850 and minimization level 1 for improved compatibility.

+++ MusicPlayerFirmware/Components/VS1053/vs1053.h
@@ -62,6 +62,6 @@ void vs1053_flush_end(void);
 
 void vs1053_start_new_track(void);
 void vs1053_send_data(const uint8_t* data, uint16_t len);
-uint32_t vs1053_get_decode_time();
+uint16_t vs1053_get_decode_time();
 
 #endif //MUSICPLAYERFIRMWARE_VS1053_H
